+++ src/lib/components/forms/asset-input/index.tsx
@@ -1,9 +1,9 @@
 import type { SystemStyleObject } from "@chakra-ui/react";
 import { Box, Button, Grid, Text } from "@chakra-ui/react";
-import { Select } from "chakra-react-select";
 
 import { CustomIcon } from "../../icon";
-import type { AssetOption } from "lib/types";
+import { SelectInput } from "../SelectInput";
+import type { AssetOption, AssetOptionValue, Option } from "lib/types";
 
 import { AssetInputFormatOptionLabel } from "./AssetInputFormatOptionLabel";
 import { AssetInputMenuList } from "./AssetInputMenuList";
@@ -16,35 +16,8 @@ interface AssetInputProps {
   setCurrencyValue: (newVal: string) => void;
   amountInput: JSX.Element;
   assetOptions: AssetOption[];
-  initialSelected?: string;
+  value: Option<AssetOption>;
 }
-
-const styles = {
-  container: (provided: SystemStyleObject) => ({
-    ...provided,
-    width: "100%",
-  }),
-  valueContainer: (provided: SystemStyleObject) => ({
-    ...provided,
-    pl: 3,
-    pr: 0,
-  }),
-  dropdownIndicator: (provided: SystemStyleObject) => ({
-    ...provided,
-    px: 2,
-    color: "gray.600",
-  }),
-  placeholder: (provided: SystemStyleObject) => ({
-    ...provided,
-    color: "gray.500",
-    fontSize: "14px",
-    whiteSpace: "nowrap",
-  }),
-  menu: (provided: SystemStyleObject) => ({
-    ...provided,
-    width: "427px",
-  }),
-};
 
 export const AssetInput = ({
   disableDelete,
@@ -52,7 +25,7 @@ export const AssetInput = ({
   setCurrencyValue,
   amountInput: AmountInput,
   assetOptions,
-  initialSelected,
+  value,
 }: AssetInputProps) => (
   <Grid
     templateColumns="140px 1fr auto"
@@ -77,35 +50,26 @@ export const AssetInput = ({
       }}
     >
       <Text className="form-label">Asset</Text>
-      <Select
-        size="lg"
+      <SelectInput<AssetOptionValue>
         placeholder="Select Asset"
         options={assetOptions}
-        defaultValue={assetOptions.find(
-          (item) => item.value === initialSelected
-        )}
-        onChange={(value) => {
-          const currentValue = (value as AssetOption).value;
-          setCurrencyValue(currentValue);
+        menuPortalTarget={document.body}
+        value={value}
+        onChange={(newValue) => {
+          if (!newValue) return;
+          setCurrencyValue(newValue.value.id);
         }}
         formatOptionLabel={AssetInputFormatOptionLabel}
-        menuPosition="fixed"
-        menuPortalTarget={document.body}
-        chakraStyles={styles}
         components={{
           Option: AssetInputOption,
           MenuList: AssetInputMenuList,
           NoOptionsMessage: AssetInputNoOptionsMessage,
         }}
-        filterOption={(
-          candidate: { label: string; value: string },
-          input: string
-        ) => {
-          if (input) {
-            return candidate.label.toLowerCase().includes(input.toLowerCase());
-          }
-
-          return true;
+        chakraStyles={{
+          menu: (provided: SystemStyleObject) => ({
+            ...provided,
+            width: "427px",
+          }),
         }}
       />
     </Box>